+++ src/test/java/devcourse/board/api/controller/PostApiControllerTest.java
@@ -1,9 +1,11 @@
-package devcourse.board.api;
+package devcourse.board.api.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import devcourse.board.domain.member.MemberService;
-import devcourse.board.domain.member.model.MemberJoinRequest;
-import devcourse.board.domain.post.PostService;
+import devcourse.board.api.model.CookieName;
+import devcourse.board.domain.member.MemberRepository;
+import devcourse.board.domain.member.model.Member;
+import devcourse.board.domain.post.PostRepository;
+import devcourse.board.domain.post.model.Post;
 import devcourse.board.domain.post.model.PostCreationRequest;
 import devcourse.board.domain.post.model.PostUpdateRequest;
 import org.junit.jupiter.api.DisplayName;
@@ -15,6 +17,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.servlet.http.Cookie;
 
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
@@ -36,18 +40,25 @@ class PostApiControllerTest {
     private ObjectMapper objectMapper;
 
     @Autowired
-    private PostService postService;
+    private MemberRepository memberRepository;
 
     @Autowired
-    private MemberService memberService;
+    private PostRepository postRepository;
 
     @Test
     @DisplayName("게시글 생성")
     void call_createPost() throws Exception {
         // given
-        Long savedMemberId = saveDummyMember("member");
+        Member member = Member.create(
+                "example@email.com",
+                "0000",
+                "member",
+                null,
+                null);
+        memberRepository.save(member);
+
         PostCreationRequest creationRequest =
-                new PostCreationRequest(savedMemberId, "title", "content");
+                new PostCreationRequest(member.getId(), "title", "content");
 
         // when & then
         mockMvc.perform(post("/posts")
@@ -70,10 +81,19 @@ class PostApiControllerTest {
     @DisplayName("게시글 단건 조회")
     void call_getPost() throws Exception {
         // given
-        Long createdPostId = createDummyPost("dummy-member", "title", "content");
+        Member member = Member.create(
+                "example@email.com",
+                "0000",
+                "member",
+                null,
+                null);
+        memberRepository.save(member);
+
+        Post post = Post.createPost(member, "post-title", "post-content");
+        postRepository.save(post);
 
         // when & then
-        mockMvc.perform(get("/posts/{postId}", createdPostId)
+        mockMvc.perform(get("/posts/{postId}", post.getId())
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andDo(print())
@@ -85,7 +105,7 @@ class PostApiControllerTest {
                                 fieldWithPath("postId").description("게시글 식별자"),
                                 fieldWithPath("title").description("게시글 제목"),
                                 fieldWithPath("content").description("게시글 내용"),
-                                fieldWithPath("createdAt").description("게시글 아이디"),
+                                fieldWithPath("createdAt").description("게시글 작성 시각"),
                                 fieldWithPath("createdBy").description("게시글 작성자 이름")
                         )));
     }
@@ -94,7 +114,16 @@ class PostApiControllerTest {
     @DisplayName("게시글 전체 조회 (페이징)")
     void call_getPosts() throws Exception {
         // given
-        createDummyPosts(20);
+        Member member = Member.create(
+                "example@email.com",
+                "0000",
+                "member",
+                null,
+                null);
+        memberRepository.save(member);
+
+        Post post = Post.createPost(member, "post-title", "post-content");
+        postRepository.save(post);
 
         // when & then
         mockMvc.perform(get("/posts")
@@ -120,14 +149,27 @@ class PostApiControllerTest {
     }
 
     @Test
-    @DisplayName("게시글 업데이트")
+    @DisplayName("회원은 자신이 작성한 게시글에 대해서만 수정할 수 있다.")
     void call_updatePost() throws Exception {
         // given
-        Long createdPostId = createDummyPost("dummy-member", "old-title", "old-content");
+        Member member = Member.create(
+                "example@email.com",
+                "0000",
+                "member",
+                null,
+                null);
+        memberRepository.save(member);
+
+        Post post = Post.createPost(member, "old-title", "old-content");
+        postRepository.save(post);
+
         PostUpdateRequest updateRequest = new PostUpdateRequest("new-title", "new-content");
 
+        Cookie idCookie = new Cookie(CookieName.MEMBER_ID.getCookieName(), String.valueOf(member.getId()));
+
         // when & then
-        mockMvc.perform(patch("/posts/{postId}", createdPostId)
+        mockMvc.perform(patch("/posts/{postId}", post.getId())
+                        .cookie(idCookie)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(updateRequest)))
                 .andExpect(status().isOk())
@@ -147,23 +189,5 @@ class PostApiControllerTest {
                                 fieldWithPath("createdAt").description("게시글 아이디"),
                                 fieldWithPath("createdBy").description("게시글 작성자 이름")
                         )));
-    }
-
-    private Long saveDummyMember(String name) {
-        return memberService.join(new MemberJoinRequest(name));
-    }
-
-    private Long createDummyPost(String memberName, String title, String content) {
-        Long savedMemberId = saveDummyMember(memberName);
-        PostCreationRequest creationRequest =
-                new PostCreationRequest(savedMemberId, title, content);
-
-        return postService.createPost(creationRequest);
-    }
-
-    private void createDummyPosts(int size) {
-        for (int i = 1; i <= size; i++) {
-            createDummyPost("member" + i, "title" + i, "content" + i);
-        }
     }
 }
